requirements.txt
python-dotenv
openai
transformers
tiktoken
llama_index
streamlit
watchdog
ntlk
llama-cpp-python